requirements.txt
langchain_openai
langchain_community
python-dotenv
streamlit
psycopg2